refactor(skills): update import paths and move skill css file to styles folder
